build(go): enforce copyloopvar + usestdlibvars; trim golangci debt (agentops-fr9 #go-lint-config) (#403)
## What (Wave 1, PR-4 of the 3.0-alignment plan — partial of
agentops-fr9)

Mechanizes more of the Go standard (`skills/standards/references/go.md`)
into the enforced `cli/.golangci.yml`, and trims pre-existing lint debt
that CI currently severity-tolerates.

**Linters enabled (both 0 findings after `--fix`):**
- `copyloopvar` — drop redundant Go 1.22+ loop-variable copies
- `usestdlibvars` — use `http.Method*` / `http.Status*` constants
instead of magic strings/ints

**Debt trimmed (already-enabled linters):**
- `errcheck`: wrapped safe `defer Close` / best-effort `os.Remove` sites
→ errcheck findings 11 → 6
- `staticcheck`: fixed S1039 (redundant `Sprintf`) + ST1005 (error
string punctuation)
- `unconvert`: removed redundant conversions in tests

## Agile decisions (not waterfall)
- `errcheck` already covers `json.Marshal/Unmarshal` (not in
`exclude-functions`) → **satisfies `agentops-tqc.3`** (no code needed;
closing it).
- **Deferred** to their fix beads, with reasons:
- `errorlint` → `agentops-bs8` (after the `%w`/`errors.Is` backlog is
cleared)
- `unconvert` + `unused` → `agentops-0l9` (`unused` ~37 backlog is risky
to auto-remove; `unconvert` has a finding in the **deprecated**
`rpi_c2_events.go` which we don't touch — needs a path exclusion first)
- The 6 residual `errcheck` findings are pre-existing atomic-write
`Close` calls where a blind `_ =` could hide data loss — left for the
careful errcheck cleanup, not silenced here.

## Test plan
- `golangci-lint v2.11.4` (CI's pinned version): `copyloopvar=0`,
`usestdlibvars=0`
- `go build ./...`, `go vet ./...`, `go test ./...` — all green (59
packages)
- `scripts/pre-push-gate.sh --fast` — passed

Bounded-context: BC5-Runtime
Evidence: golangci-lint copyloopvar=0 + usestdlibvars=0; go
build/vet/test ./... green; pre-push-gate --fast passed

+++ cli/internal/daemon/telemetry_test.go
@@ -26,7 +26,7 @@ func TestPhaseLatencyHistogram(t *testing.T) {
 		t.Fatalf("discovery histogram = %#v, want count=2 p50=10s p99=21s", discovery)
 	}
 	implementation := histograms["implementation"]
-	if implementation.Count != 1 || implementation.P50Millis != int64((3*time.Second).Milliseconds()) || implementation.P99Millis != int64((3*time.Second).Milliseconds()) {
+	if implementation.Count != 1 || implementation.P50Millis != (3*time.Second).Milliseconds() || implementation.P99Millis != (3*time.Second).Milliseconds() {
 		t.Fatalf("implementation histogram = %#v, want count=1 p50/p99=3s", implementation)
 	}
 }
